+++ CHANGELOG.rst
@@ -6,6 +6,9 @@ Changelog
 Next Release
 ------------
 
+The ``interactive_server()`` function now prepares the server to autorespond to
+the ``getFreeMonitoringStatus`` command from the mongo shell.
+
 Fix an inadvertent dependency on PyMongo.
 
 1.4.0 (2018-06-29)